+++ crates/ruma-client-api/src/media/create_content_async.rs
@@ -2,10 +2,10 @@
 //!
 //! Upload media to an MXC URI that was created with create_mxc_uri.
 
-pub mod unstable {
-    //! `/unstable/` ([spec])
+pub mod v3 {
+    //! `/v3/` ([spec])
     //!
-    //! [spec]: https://github.com/tulir/matrix-doc/blob/asynchronous_uploads/proposals/2246-asynchronous-uploads.md
+    //! [spec]: https://spec.matrix.org/latest/client-server-api/#put_matrixmediav3uploadservernamemediaid
 
     use http::header::CONTENT_TYPE;
     use ruma_common::{
@@ -19,6 +19,7 @@ pub mod unstable {
         authentication: AccessToken,
         history: {
             unstable => "/_matrix/media/unstable/fi.mau.msc2246/upload/:server_name/:media_id",
+            1.7 => "/_matrix/media/v3/upload/:server_name/:media_id",
         }
     };