+++ crates/ruma-client-api/src/media/get_content.rs
@@ -7,7 +7,6 @@ pub mod v3 {
     //!
     //! [spec]: https://spec.matrix.org/latest/client-server-api/#get_matrixmediav3downloadservernamemediaid
 
-    #[cfg(feature = "unstable-msc2246")]
     use std::time::Duration;
 
     use http::header::{CONTENT_DISPOSITION, CONTENT_TYPE};
@@ -53,16 +52,11 @@ pub mod v3 {
         /// case that the content has not yet been uploaded.
         ///
         /// The default value is 20 seconds.
-        ///
-        /// This uses the unstable prefix in
-        /// [MSC2246](https://github.com/matrix-org/matrix-spec-proposals/pull/2246).
         #[ruma_api(query)]
-        #[cfg(feature = "unstable-msc2246")]
         #[serde(
             with = "ruma_common::serde::duration::ms",
             default = "crate::media::default_download_timeout",
-            skip_serializing_if = "crate::media::is_default_download_timeout",
-            rename = "fi.mau.msc2246.max_stall_ms"
+            skip_serializing_if = "crate::media::is_default_download_timeout"
         )]
         pub timeout_ms: Duration,
 
@@ -111,7 +105,6 @@ pub mod v3 {
                 media_id,
                 server_name,
                 allow_remote: true,
-                #[cfg(feature = "unstable-msc2246")]
                 timeout_ms: crate::media::default_download_timeout(),
                 allow_redirect: false,
             }